requirements.txt
Pillow==8.2.0
pyzbar==0.1.8
opencv-python==4.5.2.54
aiohttp==3.7.4.post0
aiortc==1.2.0
